Create internal utility package
